split conmon.c into different files
conmon as it stands is pretty difficult to work with. over the years, conmon's main function has grown very large, and the conmon.c was about 1500 lines. This discourages contributors and adds maintenance burden.

instead, let's refactor conmon.c into many different files. The general structure either describes conmon lifecycle (cli process, runtime arg setup, container attach functionality, exit functions) or serves as a library (cgroup management, handling conmon oom).
