+++ meson.build
@@ -42,6 +42,28 @@ executable('conmon',
             'src/cmsg.h',
             'src/ctr_logging.c',
             'src/ctr_logging.h',
+            'src/cgroup.c',
+            'src/cgroup.h',
+            'src/cli.c',
+            'src/cli.h',
+            'src/conn_sock.c',
+            'src/conn_sock.h',
+            'src/ctr_exit.c',
+            'src/ctr_exit.h',
+            'src/ctrl.c',
+            'src/ctrl.h',
+            'src/ctr_logging.c',
+            'src/ctr_logging.h',
+            'src/ctr_stdio.c',
+            'src/ctr_stdio.h',
+            'src/globals.c',
+            'src/globals.h',
+            'src/oom.c',
+            'src/oom.h',
+            'src/parent_pipe_fd.c',
+            'src/parent_pipe_fd.h',
+            'src/runtime_args.c',
+            'src/runtime_args.h',
             'src/utils.c',
             'src/utils.h'],
            dependencies : [glib],
